wss initial
