+++ chainbench/user/__init__.py
@@ -2,7 +2,8 @@ from chainbench.user.protocol import EthBeaconUser, EvmUser, SolanaUser, StarkNe
 from chainbench.util.event import setup_event_listeners
 
 from .common import get_subclass_tasks
-from .http import HttpUser, JsonRpcUser
+from .http import HttpUser
+from .wss import WssUser
 
 # importing plugins here as all profiles depend on it
 import locust_plugins  # isort: skip  # noqa
@@ -13,8 +14,8 @@ __all__ = [
     "EthBeaconUser",
     "EvmUser",
     "HttpUser",
-    "JsonRpcUser",
     "SolanaUser",
     "StarkNetUser",
+    "WssUser",
     "get_subclass_tasks",
 ]
